refactor : 자료명 변경
- float에서 int로 변경

+++ TextRpg/Player.cs
@@ -244,7 +244,7 @@ namespace TextRpg
             }
         }
         //스킬 데미지
-        public float SkillPower(Skill skill)
+        public int SkillPower(Skill skill)
         {
             int skillPower = (int)(totalPower * skill.PowerMultiplier);
             return skillPower;
